+++ docker-compose.yml
@@ -31,6 +31,6 @@ services:
     environment:
       POSTGRES_PASSWORD: letmein
   ganache:
-    image: 'daostack/migration-experimental:0.1.1-rc.18-v0'
+    image: 'daostack/migration-experimental:0.1.1-rc.20-v0'
     ports:
       - '8545:8545'
