+++ ops/templates.json
@@ -3,67 +3,72 @@
     {
       "name": "Avatar",
       "mapping": "Avatar",
-      "start_arcVersion": "0.1.1-rc.18"
+      "start_arcVersion": "0.1.1-rc.20"
+    },
+    {
+      "name": "Vault",
+      "mapping": "Vault",
+      "start_arcVersion": "0.1.1-rc.20"
     },
     {
       "name": "Controller",
       "mapping": "Controller",
-      "start_arcVersion": "0.1.1-rc.18"
+      "start_arcVersion": "0.1.1-rc.20"
     },
     {
       "name": "DAOToken",
       "mapping": "DAOToken",
-      "start_arcVersion": "0.1.1-rc.18"
+      "start_arcVersion": "0.1.1-rc.20"
     },
     {
       "name": "Reputation",
       "mapping": "Reputation",
-      "start_arcVersion": "0.1.1-rc.18"
+      "start_arcVersion": "0.1.1-rc.20"
     },
     {
       "name": "ContributionReward",
       "mapping": "ContributionReward",
-      "start_arcVersion": "0.1.1-rc.18"
+      "start_arcVersion": "0.1.1-rc.20"
     },
     {
       "name": "ContributionRewardExt",
       "mapping": "ContributionRewardExt",
-      "start_arcVersion": "0.1.1-rc.18"
+      "start_arcVersion": "0.1.1-rc.20"
     },
     {
       "name": "Competition",
       "mapping": "Competition",
-      "start_arcVersion": "0.1.1-rc.18"
+      "start_arcVersion": "0.1.1-rc.20"
     },
     {
       "name": "GenericScheme",
       "mapping": "GenericScheme",
-      "start_arcVersion": "0.1.1-rc.18"
+      "start_arcVersion": "0.1.1-rc.20"
     },
     {
       "name": "UpgradeScheme",
       "mapping": "UpgradeScheme",
-      "start_arcVersion": "0.1.1-rc.18"
+      "start_arcVersion": "0.1.1-rc.20"
     },
     {
       "name": "JoinAndQuit",
       "mapping": "JoinAndQuit",
-      "start_arcVersion": "0.1.1-rc.18"
+      "start_arcVersion": "0.1.1-rc.20"
     },
     {
       "name": "FundingRequest",
       "mapping": "FundingRequest",
-      "start_arcVersion": "0.1.1-rc.18"
+      "start_arcVersion": "0.1.1-rc.20"
     },
     {
       "name": "SchemeFactory",
       "mapping": "SchemeFactory",
-      "start_arcVersion": "0.1.1-rc.18"
+      "start_arcVersion": "0.1.1-rc.20"
     },
     {
       "name": "SchemeRegistrar",
       "mapping": "SchemeRegistrar",
-      "start_arcVersion": "0.1.1-rc.18"
+      "start_arcVersion": "0.1.1-rc.20"
     }
   ]
 }
