+++ src/mappings/Avatar/datasource.yaml
@@ -1,8 +1,6 @@
 entities:
   - AvatarContract
-  - Vault
 eventHandlers:
-  # TODO: Add send and receive ETH with Vault support
   - event: OwnershipTransferred(indexed address,indexed address)
     handler: handleOwnershipTransferred
   - event: MetaData(string)
