+++ test/0.1.1-rc.20/SchemeRegistrar.spec.ts
