+++ test/0.1.1-rc.20/util.ts
@@ -54,7 +54,7 @@ export async function getWeb3() {
 
 export function getContractAddresses() {
   const addresses = require(`@daostack/migration-experimental/migration.json`);
-  let arcVersion = '0.1.1-rc.18';
+  let arcVersion = '0.1.1-rc.20';
 
   return {
     ...addresses.private.package[arcVersion],
@@ -74,7 +74,7 @@ export function getContractAddresses() {
 }
 
 export function getArcVersion() {
-  return '0.1.1-rc.18';
+  return '0.1.1-rc.20';
 }
 
 export function getPackageVersion() {
@@ -82,7 +82,7 @@ export function getPackageVersion() {
 }
 
 export function getOrgName() {
-  return require(`@daostack/migration-experimental/migration.json`).private.dao['0.1.1-rc.18'].name;
+  return require(`@daostack/migration-experimental/migration.json`).private.dao['0.1.1-rc.20'].name;
 }
 
 export async function getOptions(web3) {
@@ -156,9 +156,9 @@ export async function waitUntilSynced() {
 }
 
 export async function registerAdminAccountScheme(web3, addresses, opts, accounts) {
-  const Controller = require('@daostack/migration-experimental/contracts/0.1.1-rc.18/Controller.json');
-  const SchemeRegistrar = require('@daostack/migration-experimental/contracts/0.1.1-rc.18/SchemeRegistrar.json');
-  const GenesisProtocol = require('@daostack/migration-experimental/contracts/0.1.1-rc.18/GenesisProtocol.json');
+  const Controller = require('@daostack/migration-experimental/contracts/0.1.1-rc.20/Controller.json');
+  const SchemeRegistrar = require('@daostack/migration-experimental/contracts/0.1.1-rc.20/SchemeRegistrar.json');
+  const GenesisProtocol = require('@daostack/migration-experimental/contracts/0.1.1-rc.20/GenesisProtocol.json');
 
   const controller = new web3.eth.Contract(Controller.abi, addresses.Controller, opts);
   const genesisProtocol = new web3.eth.Contract(GenesisProtocol.abi, addresses.GenesisProtocol, opts);
@@ -183,8 +183,8 @@ export async function registerAdminAccountScheme(web3, addresses, opts, accounts
 }
 
 export async function prepareReputation(web3, addresses, opts, accounts) {
-  const Controller = require('@daostack/migration-experimental/contracts/0.1.1-rc.18/Controller.json');
-  const Reputation = require('@daostack/migration-experimental/contracts/0.1.1-rc.18/Reputation.json');
+  const Controller = require('@daostack/migration-experimental/contracts/0.1.1-rc.20/Controller.json');
+  const Reputation = require('@daostack/migration-experimental/contracts/0.1.1-rc.20/Reputation.json');
 
   await registerAdminAccountScheme(web3, addresses, opts, accounts);
   const controller = new web3.eth.Contract(Controller.abi, addresses.Controller, opts);
